SnapshotShipping: Added inactive resource-flag
A resource with this flag will not return any device-path for
its volumes (similar as an nvme-target). Additionally the
satellite also will skip (or even destruct) all layers above the
storage layer, so the device will not be held open by any other
layer.

When the flag is removed the satellite will also re-create
the configured layers.

+++ satellite/src/main/java/com/linbit/linstor/layer/luks/LuksLayer.java
@@ -1,6 +1,5 @@
 package com.linbit.linstor.layer.luks;
 
-import com.linbit.ImplementationError;
 import com.linbit.extproc.ExtCmdFactory;
 import com.linbit.linstor.annotation.DeviceManagerContext;
 import com.linbit.linstor.api.ApiCallRcImpl;
@@ -8,6 +7,7 @@ import com.linbit.linstor.core.devmgr.DeviceHandler;
 import com.linbit.linstor.core.devmgr.exceptions.ResourceException;
 import com.linbit.linstor.core.devmgr.exceptions.VolumeException;
 import com.linbit.linstor.core.objects.Resource;
+import com.linbit.linstor.core.objects.Resource.Flags;
 import com.linbit.linstor.core.objects.Snapshot;
 import com.linbit.linstor.core.objects.Volume;
 import com.linbit.linstor.dbdrivers.DatabaseException;
@@ -18,6 +18,7 @@ import com.linbit.linstor.logging.ErrorReporter;
 import com.linbit.linstor.propscon.Props;
 import com.linbit.linstor.security.AccessContext;
 import com.linbit.linstor.security.AccessDeniedException;
+import com.linbit.linstor.stateflags.StateFlags;
 import com.linbit.linstor.storage.StorageException;
 import com.linbit.linstor.storage.data.adapter.luks.LuksRscData;
 import com.linbit.linstor.storage.data.adapter.luks.LuksVlmData;
@@ -30,9 +31,7 @@ import javax.inject.Provider;
 import javax.inject.Singleton;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 @Singleton
 public class LuksLayer implements DeviceLayer
@@ -83,20 +82,22 @@ public class LuksLayer implements DeviceLayer
     @Override
     public void resourceFinished(AbsRscLayerObject<Resource> layerDataRef) throws AccessDeniedException
     {
-        if (layerDataRef.getAbsResource().getStateFlags().isSet(sysCtx, Resource.Flags.DELETE))
+        StateFlags<Flags> rscFlags = layerDataRef.getAbsResource().getStateFlags();
+        if (rscFlags.isSet(sysCtx, Resource.Flags.DELETE))
         {
             resourceProcessorProvider.get().sendResourceDeletedEvent(layerDataRef);
         }
         else
         {
+            boolean isActive = rscFlags.isSet(sysCtx, Resource.Flags.INACTIVE);
             resourceProcessorProvider.get().sendResourceCreatedEvent(
                 layerDataRef,
                 new UsageState(
-                    true,
+                    isActive,
                     // we could check here if one of your LuksVlms is open - but this method is only called
                     // right after the creation, where nothing can be in use now.
                     false,
-                    true
+                    isActive
                 )
             );
         }
@@ -152,26 +153,9 @@ public class LuksLayer implements DeviceLayer
     {
         LayerProcessResult ret;
         LuksRscData<Resource> luksRscData = (LuksRscData<Resource>) rscData;
-        boolean deleteRsc = luksRscData.getAbsResource().getStateFlags().isSet(sysCtx, Resource.Flags.DELETE);
-
-        Map<Boolean, List<LuksVlmData<Resource>>> groupedByDeleteFlag =
-            luksRscData.getVlmLayerObjects().values().stream().collect(
-                Collectors.partitioningBy(luksVlmData ->
-                {
-                    boolean isDeleted;
-                    try
-                    {
-                        isDeleted = deleteRsc ||
-                            ((Volume) luksVlmData.getVolume()).getFlags().isSet(sysCtx, Volume.Flags.DELETE);
-                    }
-                    catch (AccessDeniedException exc)
-                    {
-                        throw new ImplementationError(exc);
-                    }
-                    return isDeleted;
-                }
-            )
-        );
+        StateFlags<Flags> flags = luksRscData.getAbsResource().getStateFlags();
+        boolean deleteRsc = flags.isSet(sysCtx, Resource.Flags.DELETE);
+        boolean deactivateRsc = flags.isSet(sysCtx, Resource.Flags.INACTIVE);
 
         boolean allVolumeKeysDecrypted = true;
         for (LuksVlmData<Resource> vlmData : luksRscData.getVlmLayerObjects().values())
@@ -185,41 +169,53 @@ public class LuksLayer implements DeviceLayer
 
         if (allVolumeKeysDecrypted) // otherwise do not even process children.
         {
-            for (LuksVlmData<Resource> vlmData : groupedByDeleteFlag.get(true))
+            for (LuksVlmData<Resource> vlmData : luksRscData.getVlmLayerObjects().values())
             {
-                String identifier = getIdentifier(vlmData);
-
-                if (cryptSetup.isOpen(identifier))
-                {
-                    cryptSetup.closeLuksDevice(identifier);
-                    vlmData.setOpened(false);
-                    vlmData.setExists(false);
-                    cryptSetup.deleteHeaders(vlmData.getBackingDevice());
-                }
-            }
-
-            // shrink if necessary
-            for (LuksVlmData<Resource> vlmData : groupedByDeleteFlag.get(false))
-            {
-                if (
-                    ((Volume) vlmData.getVolume()).getFlags().isSet(sysCtx, Volume.Flags.RESIZE) &&
-                        vlmData.getSizeState().equals(Size.TOO_LARGE)
-                )
+                boolean deleteVlm = deleteRsc ||
+                    ((Volume) vlmData.getVolume()).getFlags().isSet(sysCtx, Volume.Flags.DELETE);
+                if (deleteVlm || deactivateRsc)
                 {
                     String identifier = getIdentifier(vlmData);
 
-                    boolean isOpen = cryptSetup.isOpen(identifier);
-
-                    vlmData.setBackingDevice(vlmData.getSingleChild().getDevicePath());
-
-                    boolean alreadyLuks = cryptSetup.hasLuksFormat(vlmData);
-
-                    if (isOpen && alreadyLuks)
+                    if (cryptSetup.isOpen(identifier))
                     {
-                        cryptSetup.shrink(vlmData);
-                        // do not set Size.AS_EXPECTED as we will need to grow as much as we can
-                        // once the layers below us are done shrinking
-                        // vlmData.setSizeState(Size.AS_EXPECTED);
+                        cryptSetup.closeLuksDevice(identifier);
+                        vlmData.setOpened(false);
+                        vlmData.setExists(false);
+
+                        if (deleteVlm)
+                        {
+                            cryptSetup.deleteHeaders(vlmData.getBackingDevice());
+                        }
+                        else
+                        {
+                            vlmData.setDevicePath(null);
+                        }
+                    }
+                }
+                else
+                {
+                    // shrink if necessary
+                    if (
+                        ((Volume) vlmData.getVolume()).getFlags().isSet(sysCtx, Volume.Flags.RESIZE) &&
+                            vlmData.getSizeState().equals(Size.TOO_LARGE)
+                    )
+                    {
+                        String identifier = getIdentifier(vlmData);
+
+                        boolean isOpen = cryptSetup.isOpen(identifier);
+
+                        vlmData.setBackingDevice(vlmData.getSingleChild().getDevicePath());
+
+                        boolean alreadyLuks = cryptSetup.hasLuksFormat(vlmData);
+
+                        if (isOpen && alreadyLuks)
+                        {
+                            cryptSetup.shrink(vlmData);
+                            // do not set Size.AS_EXPECTED as we will need to grow as much as we can
+                            // once the layers below us are done shrinking
+                            // vlmData.setSizeState(Size.AS_EXPECTED);
+                        }
                     }
                 }
             }
@@ -232,69 +228,74 @@ public class LuksLayer implements DeviceLayer
 
             if (processResult == LayerProcessResult.SUCCESS)
             {
-                for (LuksVlmData<Resource> vlmData : groupedByDeleteFlag.get(false))
+                for (LuksVlmData<Resource> vlmData : luksRscData.getVlmLayerObjects().values())
                 {
-                    String identifier = getIdentifier(vlmData);
-
-                    boolean isOpen = cryptSetup.isOpen(identifier);
-
-                    vlmData.setBackingDevice(vlmData.getSingleChild().getDevicePath());
-
-                    boolean alreadyLuks = cryptSetup.hasLuksFormat(vlmData);
-
-                    if (!alreadyLuks)
+                    boolean deleteVlm = deleteRsc ||
+                        ((Volume) vlmData.getVolume()).getFlags().isSet(sysCtx, Volume.Flags.DELETE);
+                    if (!deleteVlm && !deactivateRsc)
                     {
-                        String providedDev = cryptSetup.createLuksDevice(
-                            vlmData.getBackingDevice(),
-                            vlmData.getDecryptedPassword(),
-                            identifier
+                        String identifier = getIdentifier(vlmData);
+
+                        boolean isOpen = cryptSetup.isOpen(identifier);
+
+                        vlmData.setBackingDevice(vlmData.getSingleChild().getDevicePath());
+
+                        boolean alreadyLuks = cryptSetup.hasLuksFormat(vlmData);
+
+                        if (!alreadyLuks)
+                        {
+                            String providedDev = cryptSetup.createLuksDevice(
+                                vlmData.getBackingDevice(),
+                                vlmData.getDecryptedPassword(),
+                                identifier
+                            );
+                            vlmData.setDevicePath(providedDev);
+
+                            // prevent unnecessary resize when we just created the device
+                            vlmData.setSizeState(Size.AS_EXPECTED);
+                        }
+                        else
+                        {
+                            /*
+                             * TODO: this step should not be necessary
+                             * currently it is, because LayeredResourceHelper re-creates the LuksVlmData
+                             * in every iteration. Once those data live longer (or are restored from props)
+                             * the next command can be removed
+                             */
+                            vlmData.setDevicePath(cryptSetup.getLuksVolumePath(identifier));
+                        }
+
+                        if (!isOpen)
+                        {
+                            cryptSetup
+                                .openLuksDevice(vlmData.getBackingDevice(), identifier, vlmData.getDecryptedPassword());
+                        }
+
+                        if (
+                            ((Volume) vlmData.getVolume()).getFlags().isSet(sysCtx, Volume.Flags.RESIZE) &&
+                                !vlmData.getSizeState().equals(Size.AS_EXPECTED)
+                        )
+                        {
+                            cryptSetup.grow(vlmData);
+                        }
+
+                        vlmData.setAllocatedSize(
+                            Commands.getBlockSizeInKib(
+                                extCmdFactory.create(),
+                                vlmData.getBackingDevice()
+                            )
                         );
-                        vlmData.setDevicePath(providedDev);
-
-                        // prevent unnecessary resize when we just created the device
+                        vlmData.setUsableSize(
+                            Commands.getBlockSizeInKib(
+                                extCmdFactory.create(),
+                                vlmData.getDevicePath()
+                            )
+                        );
                         vlmData.setSizeState(Size.AS_EXPECTED);
-                    }
-                    else
-                    {
-                        /*
-                         * TODO: this step should not be necessary
-                         * currently it is, because LayeredResourceHelper re-creates the LuksVlmData
-                         * in every iteration. Once those data live longer (or are restored from props)
-                         * the next command can be removed
-                         */
-                        vlmData.setDevicePath(cryptSetup.getLuksVolumePath(identifier));
-                    }
 
-                    if (!isOpen)
-                    {
-                        cryptSetup
-                            .openLuksDevice(vlmData.getBackingDevice(), identifier, vlmData.getDecryptedPassword());
+                        vlmData.setOpened(true);
+                        vlmData.setFailed(false);
                     }
-
-                    if (
-                        ((Volume) vlmData.getVolume()).getFlags().isSet(sysCtx, Volume.Flags.RESIZE) &&
-                            !vlmData.getSizeState().equals(Size.AS_EXPECTED)
-                    )
-                    {
-                        cryptSetup.grow(vlmData);
-                    }
-
-                    vlmData.setAllocatedSize(
-                        Commands.getBlockSizeInKib(
-                            extCmdFactory.create(),
-                            vlmData.getBackingDevice()
-                        )
-                    );
-                    vlmData.setUsableSize(
-                        Commands.getBlockSizeInKib(
-                            extCmdFactory.create(),
-                            vlmData.getDevicePath()
-                        )
-                    );
-                    vlmData.setSizeState(Size.AS_EXPECTED);
-
-                    vlmData.setOpened(true);
-                    vlmData.setFailed(false);
                 }
                 ret = LayerProcessResult.SUCCESS;
             }
@@ -305,16 +306,18 @@ public class LuksLayer implements DeviceLayer
         }
         else
         {
-            errorReporter.logWarning(
-                "Luks layer cannot process resource '%s' because some volumes " +
-                    "are missing the decrypted key. Is the master key set?",
-                luksRscData.getSuffixedResourceName()
-            );
             for (LuksVlmData<Resource> vlmData : luksRscData.getVlmLayerObjects().values())
             {
                 vlmData.setFailed(true);
             }
-            ret = LayerProcessResult.NO_DEVICES_PROVIDED;
+            throw new AbortLayerProcessingException(
+                luksRscData,
+                String.format(
+                    "Luks layer cannot process resource '%s' because some volumes " +
+                        "are missing the decrypted key. Is the master key set?",
+                    luksRscData.getSuffixedResourceName()
+                )
+            );
         }
         return ret;
     }
